requirements.txt
discord.py==1.7.3
openpyxl==3.0.7
pytz==2021.1
python-dotenv==0.19.0